refactor: move hooks/ directory to src/hooks/
- Move all files from hooks/ to src/hooks/ (17 files)
- Update all imports from './hooks/' or '../hooks/' to '@/hooks/' across codebase
- Fix useSnippetLoader.ts import path for snippets/default-snippets.json (now ../../snippets)
- Fix test file imports in hooks/*.test.ts to use ../../test/mocks/
- Fix worker path in useRenpyAnalysis.ts (now ../../workers)
- Verification: All 481 tests passing, build succeeds in 2.51s
