+++ App.tsx
@@ -25,8 +25,8 @@ import ScreenLayoutComposer from './components/ScreenLayoutComposer';
 import MarkdownPreviewView from './components/MarkdownPreviewView';
 import DiagnosticsPanel from './components/DiagnosticsPanel';
 import WarpVariablesModal from './components/WarpVariablesModal';
-import { useDiagnostics, migratePunchlistToTasks } from './hooks/useDiagnostics';
-import { useDebounce } from './hooks/useDebounce';
+import { useDiagnostics, migratePunchlistToTasks } from '@/hooks/useDiagnostics';
+import { useDebounce } from '@/hooks/useDebounce';
 import TabContextMenu from './components/TabContextMenu';
 import Sash from './components/Sash';
 import StatusBar from './components/StatusBar';
@@ -40,10 +40,10 @@ import { SearchProvider } from './contexts/SearchContext';
 import StatsView from './components/StatsView';
 import TranslationDashboard from './components/TranslationDashboard';
 import GoToLabelModal, { GoToLabelItem } from './components/GoToLabelModal';
-import { useRenpyAnalysis, performRouteAnalysis } from './hooks/useRenpyAnalysis';
-import { useHistory } from './hooks/useHistory';
-import { useProjectColorScan } from './hooks/useProjectColorScan';
-import { usePerformanceMetrics } from './hooks/usePerformanceMetrics';
+import { useRenpyAnalysis, performRouteAnalysis } from '@/hooks/useRenpyAnalysis';
+import { useHistory } from '@/hooks/useHistory';
+import { useProjectColorScan } from '@/hooks/useProjectColorScan';
+import { usePerformanceMetrics } from '@/hooks/usePerformanceMetrics';
 import { createId } from '@/lib/createId';
 import { formatErrorMessage } from '@/lib/formatErrorMessage';
 import {
