+++ components/MenuConstructorModal.tsx
@@ -8,7 +8,7 @@
 import React, { useState, useEffect, useRef } from 'react';
 import { createPortal } from 'react-dom';
 import type { MenuChoice, MenuTemplate } from '@/types';
-import { useModalAccessibility } from '../hooks/useModalAccessibility';
+import { useModalAccessibility } from '@/hooks/useModalAccessibility';
 import { createId } from '@/lib/createId';
 import * as monaco from 'monaco-editor';
 import CodeActionButtons from './CodeActionButtons';
