+++ components/NewProjectWizardModal.tsx
@@ -11,7 +11,7 @@ import React, { useState, useCallback, useEffect } from 'react';
 import { logger } from '../lib/logger';
 import { createPortal } from 'react-dom';
 import { CreateProjectOptions } from '@/types';
-import { useModalAccessibility } from '../hooks/useModalAccessibility';
+import { useModalAccessibility } from '@/hooks/useModalAccessibility';
 import ColorDropTarget from './ColorDropTarget';
 
 interface NewProjectWizardModalProps {
