+++ components/SettingsModal.tsx
@@ -7,7 +7,7 @@
  */
 
 import React from 'react';
-import { useModalAccessibility } from '../hooks/useModalAccessibility';
+import { useModalAccessibility } from '@/hooks/useModalAccessibility';
 import type { Theme, IdeSettings, MouseGestureSettings, CanvasPanGesture } from '@/types';
 
 interface SettingsModalProps {
