+++ components/SnippetManager.tsx
@@ -14,7 +14,7 @@ import React from 'react';
 import type { UserSnippet } from '@/types';
 import CopyButton from './CopyButton';
 import SnippetGridView from './SnippetGridView';
-import { useSnippetLoader } from '../hooks/useSnippetLoader';
+import { useSnippetLoader } from '@/hooks/useSnippetLoader';
 
 interface SnippetManagerProps {
     userSnippets?: UserSnippet[];
