+++ components/WarpVariablesModal.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useMemo, useRef, useState } from 'react';
 import { createPortal } from 'react-dom';
-import { useModalAccessibility } from '../hooks/useModalAccessibility';
+import { useModalAccessibility } from '@/hooks/useModalAccessibility';
 import type { WarpVariableDraft } from '@/lib/warpAfterWarp';
 
 interface WarpVariablesModalProps {
