+++ src/hooks/useDiagnostics.ts
