+++ src/hooks/useFileSystemManager.ts
