+++ src/hooks/usePerformanceMetrics.ts
