+++ src/hooks/useProjectColorScan.ts
