+++ workers/renpyAnalysis.worker.ts
@@ -8,8 +8,8 @@
  * immediately without re-running the full analysis pass.
  */
 
-import { performRenpyAnalysis, performRouteAnalysis } from '../hooks/useRenpyAnalysis';
-import type { AnalysisBlock } from '../hooks/useRenpyAnalysis';
+import { performRenpyAnalysis, performRouteAnalysis } from '@/hooks/useRenpyAnalysis';
+import type { AnalysisBlock } from '@/hooks/useRenpyAnalysis';
 import type { RenpyAnalysisResult } from '@/types';
 import { formatErrorMessage } from '@/lib/formatErrorMessage';
 import { performTranslationAnalysis } from '@/lib/renpyTranslationParser';
